+++ js/IN_createRaffle.js
@@ -1,1 +1,0 @@
-Inject('createRaffle.js');